Move around some namespaces.
