+++ osu.Game/Screens/Tournament/Teams/Team.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
-namespace osu.Game.Screens.Tournament
+namespace osu.Game.Screens.Tournament.Teams
 {
     public class Team
     {
